# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Examination Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, completed, all]
        - name: subject
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Exam list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamListResponse"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExamBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exams/{id}/questions:
    get:
      operationId: getExamQuestions
      tags: [exams]
      summary: Get questions assigned to an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamQuestion"
    post:
      operationId: assignQuestionsToExam
      tags: [exams]
      summary: Assign questions to exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignQuestionsBody"
      responses:
        "200":
          description: Questions assigned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamQuestion"
  /exams/{id}/auto-generate:
    post:
      operationId: autoGenerateQuestionPaper
      tags: [exams]
      summary: Auto-generate question paper from question bank
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoGenerateBody"
      responses:
        "200":
          description: Questions auto-assigned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamQuestion"
  /exams/{id}/eligible-students:
    get:
      operationId: getEligibleStudents
      tags: [exams]
      summary: Get students eligible for exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eligible students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

  # ── Questions ──
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        rollNo:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        course:
          type: string
        year:
          type: integer
        photoUrl:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, rollNo, name, email, phone, course, year, createdAt]
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        examName:
          type: string
        subject:
          type: string
        examDate:
          type: string
          format: date-time
        startTime:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        examType:
          type: string
          enum: [mcq]
        eligibleCourses:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled, ongoing, completed, cancelled]
        resultsPublished:
          type: boolean
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, examName, subject, examDate, startTime, duration, totalMarks, passingMarks, examType, status, resultsPublished, createdAt]
    ExamDetail:
      type: object
      properties:
        id:
          type: integer
        examName:
          type: string
        subject:
          type: string
        examDate:
          type: string
          format: date-time
        startTime:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        examType:
          type: string
          enum: [mcq]
        eligibleCourses:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled, ongoing, completed, cancelled]
        resultsPublished:
          type: boolean
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        questionCount:
          type: integer
        studentCount:
          type: integer
        attendanceCount:
          type: integer
      required: [id, examName, subject, examDate, startTime, duration, totalMarks, passingMarks, examType, status, resultsPublished, createdAt, questionCount, studentCount, attendanceCount]
    CreateExamBody:
      type: object
      properties:
        examName:
          type: string
        subject:
          type: string
        examDate:
          type: string
          format: date-time
        startTime:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        examType:
          type: string
          enum: [mcq]
        eligibleCourses:
          type: array
          items:
            type: string
      required: [examName, subject, examDate, startTime, duration, totalMarks, passingMarks]
    UpdateExamBody:
      type: object
      properties:
        examName:
          type: string
        subject:
          type: string
        examDate:
          type: string
          format: date-time
        startTime:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        examType:
          type: string
          enum: [mcq]
        eligibleCourses:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled, ongoing, completed, cancelled]
    ExamListResponse:
      type: object
      properties:
        exams:
          type: array
          items:
            $ref: "#/components/schemas/Exam"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [exams, total, page, totalPages]
    Question:
      type: object
      properties:
        id:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq]
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: ["string", "null"]
        marks:
          type: integer
        markingScheme:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, questionText, questionType, subject, topic, difficulty, marks, createdAt]
...
    ExamQuestion:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        questionId:
          type: integer
        questionOrder:
          type: integer
        question:
          $ref: "#/components/schemas/Question"
      required: [id, examId, questionId, questionOrder]
    AssignQuestionsBody:
      type: object
      properties:
        questionIds:
          type: array
          items:
            type: integer
      required: [questionIds]
    AutoGenerateBody:
      type: object
      properties:
        subject:
          type: string
        count:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
      required: [subject, count]
...